Add db subpackage
Management of a local sqlite database which is to house data on the
user's SimpleLogin objects, like aliases and mailboxes. This will
enable users to search for and manipulate their SimpleLogin data more
quickly. The local db will not, however, be intended as a complete
replacement of API queries, since objects and object states can and
likely will be modified, added, and removed by other applications,
       like the SimpleLogin browser extensions and phone apps. The CLI
       will therefore need to provide some means of syncing user data
       as well as to clarify which functionality relates specifically
to the local db versus SimpleLogin's db.