Add SimpleLoginSession subclass
Modestly extend the usual SQLAlchemy Session to abstract merging
functionality into a `.upsert()` method, which has the effect of
creating new objects or modifying existing ones based on the primary
key of the object passed in. Make this subclass the default used by
the DatabaseAccessLayer so that upserts are clean and concise.