Add LazyGroup class
First real step in addressing the app's very slow runtime, define a
subclass of Click Group that lazily-loads its commands as they're
needed, rather than loading all commands together.

Also provide a small utility function for users to define the location
of subcommands comveniently. It's perhaps not an ideal means of
providing the class with that info, but it should be sufficiently
both generic and concise enough for regular use.