Defer command implementations
At the risk of further complicating the project's command structure,
move all concrete group/command implementations to private modules,
and modify all public group/command callbacks to import and call those
implementations locally.  The effect is that the runtime of `--help`
and similar commands is vastly reduced (by nearly 10x in some cases),
because some of those imports are slow, and many are unnecessary for
these types of eager-option callbacks.

Also remove all actions from the main cli group's callback function,
deferring config, database, and other initialization until the commands
that require them are actually invoked. This means more boilerplate
for many command implementations, but is necessary for faster startup.
Further, the objects which were previously passed around via Click's
context--the db, sl, and config objects--are now to be loaded via the
`simplelogincmd.cli.util.init_*` functions, so that commands have finer
control over what they load, rather than naively loading everything
all the time.

The means by which users are checked and potentially automatically
"logged in" also has changed. Commands no longer require the
`authenticate` decorator, which has been removed. Instead, an API key
is saught and added to the SimpleLogin client whenever one is created
via the new `util.init_sl` function. The function also handles
prompting for email/password/OTP if no API key is found. SO, in short,
          all a command need do in order to invoke the previous auto-
          login functionality is simpley to call that function.

The convention for organizing commands is now thus. Each top-level
group lives in a public module directly under the
`/simplelogincmd/cli/commands/` directory. The module and the group
should have the same name. The group's concrete implementation, if any,
is defined in a private module in the same location. The group's
subgroups and subcommands are located in a subdirectory under
`/simplelogincmd/cli/commands/` that shares the group's name with
"_commands" appended. As with the top-level group, subgroup/subcommand
definitions are provided in a public module named after the subgroup/
subcommand, and implementations reside in private modules at the same
level. This pattern can repeat as deeply as groups require.

Below is a brief visual example, because the above explanation probably
makes little sense:

commands/
    alias_commands/
        contact_commands/
            _create.py
            create.py
        _get.py
        _list.py
        get.py
        list.py
    _alias.py
    alias.py